AMD-694-last-fix [frontend] refactor: add styledBoldText to file
