+++ frontend/src/containers/style.tsx
@@ -1,0 +1,7 @@
+import { styled } from '@mui/material/styles';
+import { theme } from '@src/theme';
+
+export const StyledBoldText = styled('span')({
+  color: theme.main.secondColor,
+  fontWeight: 600,
+});
