+++ frontend/src/hooks/reportMapper/reworkMapper.tsx
@@ -1,7 +1,7 @@
 import { ReportDataWithTwoColumns } from '@src/hooks/reportMapper/reportUIDataStructure';
 import { REWORK_BOARD_STATUS, REWORK_TIME_MAPPING } from '@src/constants/resources';
-import { StyledBoldText } from '@src/containers/ConfigStep/TimeoutAlert/style';
 import { ReworkTimeResponse } from '@src/clients/report/dto/response';
+import { StyledBoldText } from '@src/containers/style';
 import { ReactNode } from 'react';
 
 const getUnit = (value: string) => {
